Standardize KDEModel algorithm mode related methods
All methods in KDEModel have a similar interface to set up KDE model
parameters, except for the mode of the algorithm. This patch refactors
KDEModel to provide a similar interface for modes.

Notice that the public API and serialization is changed, but this patch
is aimed at the release of a new mlpack major version.